requirements.txt
fastapi==0.116.1
pinecone==7.3.0
pydantic==2.11.7
PyPDF2==3.0.1
python-dotenv==1.1.1
Requests==2.32.4
tiktoken==0.9.0
